reconfigured reinstall job

+++ bin/haproxy_get_ssl.py
@@ -4,18 +4,29 @@ from jinja2 import Template
 
 r_server = redis.StrictRedis('127.0.0.1', db=2)
 check = r_server.get("need_CSR")
-check = "1"
-if check == "1":
-    i_key = "owner-info"
-    data=json.loads (r_server.get(i_key))
+if check == None:
+    print "need run registration job" 
+
+if len(sys.argv) > 1:
+    need_haproxy=sys.argv[1]
+else:
+    need_haproxy='no'
+
+i_key = "owner-info"
+json_data = r_server.get(i_key)
+if json_data != None:
+    data=json.loads (json_data)
     email = data['Email']
     hostname = data['Hostname']
-    frontend_conf = ""
-    backend_conf = ""
-    acl = ""
-    redirect = ""
+frontend_conf = ""
+backend_conf = ""
+acl = ""
+redirect = ""
 
+if hostname == None:
+    hostname = os.environ["HOSTNAME"]
 
+if check == "1":
 #### get certificate
     if os.environ["ENV"] == 'AWS':
         os.system("mkdir -p /opt/certs/letsencrypt")
@@ -26,9 +37,12 @@ if check == "1":
         os.system("rm -rf /opt/certs/letsencrypt")
         r_server.set("need_CSR", "0")
         r_server.bgsave
-
+else:
+    print "Don't need new certificate"
 
 ### reconfigure haproxy
+if (check == "1") or (need_haproxy == "yes") :
+
     app_key="apps"
     data_apps=r_server.get(app_key)
     os.system("rm -rf /opt/haproxy/haproxy.cfg")
@@ -66,4 +80,4 @@ if check == "1":
     os.system("haproxy -f /opt/haproxy/haproxy.cfg -p /var/run/haproxy.pid -sf $(cat /var/run/haproxy.pid)")
     
 else:
-    print "Don't need new certificate"
+    print "Don't need reconfigure HAproxy"
